requirements.txt
boto3==1.12.0
botocore==1.15.0
requests==2.22.0
PyJWT==1.7.1
